Added diko-init to mint foundation and founder tokens at start, and keep track of distribution.
